+++ map/google_foss/lib/google_map_availability.dart
@@ -1,0 +1,14 @@
+import 'package:map_core/map_availability.dart';
+
+class GoogleMapAvailability extends MapAvailability {
+
+  @override
+  bool isAvailable() {
+    return false;
+  }
+
+  @override
+  Future<bool> isDefault() {
+    return Future.value(false);
+  }
+}