+++ sponsor/data_closed/lib/di/sponsor_module_factory.dart
@@ -1,8 +1,0 @@
-
-import 'package:sponsor_core/repository/membership_repository.dart';
-
-// class SponsorModuleFactory {
-//   static MembershipRepository createMembershipRepository() {
-//     return StoreMembershipRepository();
-//   }
-// }