+++ sponsor/ui/lib/ui/setting/cubit/sponsor_setting_item_cubit.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sponsor_core/repository/membership_repository.dart';
-import 'package:sponsor_data_closed/repository/store_membership_repository.dart';
+import 'package:sponsor_data/di/sponsor_data_module_factory.dart';
 import 'package:sponsor_ui/ui/setting/cubit/sponsor_setting_item_state.dart';
 
 class SponsorSettingItemCubit extends Cubit<SponsorSettingItemState> {
   final MembershipRepository _membershipRepository =
-      StoreMembershipRepository();
+      SponsorDataModuleFactory.createMembershipRepository();
 
   SponsorSettingItemCubit() : super(LoadingSponsorSettingItemState()) {
     _fetchMembership();